requirements.txt
pandas
streamlit
python-docx
docx2pdf
plotly
